requirements.txt
tornado==6.0.4
pika==1.1.0
python-decouple==3.3
